+++ packages/commerce-sdk-react/src/hooks/useCustomerId.ts
@@ -4,10 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import useAuthContext from './useAuthContext'
-import useLocalStorage from './useLocalStorage'
-import useConfig from './useConfig'
-import {onClient} from '../utils'
+import useAuthDataValue from './useAuthDataValue'
 
 /**
  * Hook that returns the customer ID.
@@ -15,18 +12,6 @@ import {onClient} from '../utils'
  * @group Helpers
  * @category Shopper Authentication
  */
-const useCustomerId = (): string | null => {
-    if (onClient()) {
-        // `onClient` is a constant, so the hooks will always have the same execution order,
-        // despite technically being inside a conditional.
-        // eslint-disable-next-line react-hooks/rules-of-hooks
-        const config = useConfig()
-        // eslint-disable-next-line react-hooks/rules-of-hooks
-        return useLocalStorage(`customer_id_${config.siteId}`)
-    }
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    const auth = useAuthContext()
-    return auth.get('customer_id')
-}
+const useCustomerId = (): string | null => useAuthDataValue('customer_id')
 
 export default useCustomerId